Split all imitation algorithm wrappers into separate files
