+++ src/rl_framework/agent/imitation/__init__.py
@@ -1,6 +1,2 @@
 from .episode_sequence import EpisodeSequence
-from .imitation import (
-    IMITATION_ALGORITHM_WRAPPER_REGISTRY,
-    AlgorithmWrapper,
-    ImitationAgent,
-)
+from .imitation import IMITATION_ALGORITHM_WRAPPER_REGISTRY, ImitationAgent
