+++ src/rl_framework/agent/imitation/imitation/__init__.py
@@ -1,2 +1,2 @@
+from .algorithms import AlgorithmWrapper
 from .imitation import IMITATION_ALGORITHM_WRAPPER_REGISTRY, ImitationAgent
-from .imitation_algorithm_wrappers import AlgorithmWrapper
