+++ src/rl_framework/agent/imitation/imitation/algorithms/__init__.py
@@ -1,0 +1,6 @@
+from .airl import AIRLAlgorithmWrapper
+from .algorithm_wrapper import AlgorithmWrapper
+from .bc import BCAlgorithmWrapper
+from .density import DensityAlgorithmWrapper
+from .gail import GAILAlgorithmWrapper
+from .sqil import SQILAlgorithmWrapper
